fix(server): serve a workspace's total file size as an integer
SUM over the bigint file_size column returns numeric, which psycopg hands
back as a Decimal, so the routes reporting a workspace's backed-up size
served a JSON string instead of the number the web types declare. The query
now casts the sum back to bigint.

+++ src/server/database/workspace_file.py
@@ -660,9 +660,11 @@ async def get_workspace_total_size(
     try:
 
         async def _execute(cur):
+            # SUM over a bigint is numeric, which arrives as a Decimal and
+            # serializes as a string.
             await cur.execute(
                 """
-                SELECT COALESCE(SUM(file_size), 0) AS total_size
+                SELECT COALESCE(SUM(file_size), 0)::bigint AS total_size
                 FROM workspace_files
                 WHERE workspace_id = %s
                 """,
